Add epub navigator fragment
