+++ r2-navigator/src/main/java/org/readium/r2/navigator/epub/R2EpubActivity.kt
@@ -9,9 +9,7 @@
 
 package org.readium.r2.navigator.epub
 
-import android.app.Activity
 import android.content.Context
-import android.content.Intent
 import android.content.SharedPreferences
 import android.graphics.Color
 import android.graphics.Rect
@@ -20,202 +18,37 @@ import android.util.DisplayMetrics
 import android.view.ActionMode
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.net.toUri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.viewpager.widget.ViewPager
 import kotlinx.coroutines.*
 import org.json.JSONException
 import org.json.JSONObject
 import org.readium.r2.navigator.*
-import org.readium.r2.navigator.extensions.layoutDirectionIsRTL
-import org.readium.r2.navigator.extensions.positionsByResource
 import org.readium.r2.navigator.pager.R2EpubPageFragment
 import org.readium.r2.navigator.pager.R2PagerAdapter
 import org.readium.r2.navigator.pager.R2ViewPager
-import org.readium.r2.shared.COLUMN_COUNT_REF
-import org.readium.r2.shared.SCROLL_REF
+import org.readium.r2.shared.FragmentNavigator
 import org.readium.r2.shared.extensions.destroyPublication
 import org.readium.r2.shared.extensions.getPublication
-import org.readium.r2.shared.getAbsolute
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.publication.ReadingProgression
-import org.readium.r2.shared.publication.epub.EpubLayout
-import org.readium.r2.shared.publication.presentation.presentation
-import org.readium.r2.shared.publication.services.positions
-import java.net.URI
 import kotlin.coroutines.CoroutineContext
-import kotlin.math.ceil
 
 
-open class R2EpubActivity : AppCompatActivity(), IR2Activity, IR2Selectable, IR2Highlightable, IR2TTS, CoroutineScope, VisualNavigator {
-
-    override fun progressionDidChange(progression: Double) {
-        notifyCurrentLocation()
-    }
+open class R2EpubActivity: AppCompatActivity(), IR2Activity, IR2Selectable, IR2Highlightable, IR2TTS, CoroutineScope, VisualNavigator {
 
     /**
      * Locator waiting to be loaded in the navigator.
      */
     internal var pendingLocator: Locator? = null
 
-    override fun go(locator: Locator, animated: Boolean, completion: () -> Unit): Boolean {
-        pendingLocator = locator
-
-        // href is the link to the page in the toc
-        var href = locator.href
-
-        if (href.indexOf("#") > 0) {
-            href = href.substring(0, href.indexOf("#"))
-        }
-
-        fun setCurrent(resources: ArrayList<*>) {
-            for (resource in resources) {
-                if (resource is Pair<*, *>) {
-                    resource as Pair<Int, String>
-                    if (resource.second.endsWith(href)) {
-                        if (resourcePager.currentItem == resource.first) {
-                            // reload webview if it has an anchor
-                            locator.locations.fragments.firstOrNull()?.let { fragment ->
-
-                                val fragments = fragment.split(",").associate {
-                                    val (left, right) = it.split("=")
-                                    left to right.toInt()
-                                }
-                                //            val id = fragments.getValue("id")
-                                if (fragments.isEmpty()) {
-                                    var anchor = fragment
-                                    if (!anchor.startsWith("#")) {
-                                        anchor = "#$anchor"
-                                    }
-                                    val goto = resource.second + anchor
-                                    currentFragment?.webView?.loadUrl(goto)
-                                } else {
-                                    currentFragment?.webView?.loadUrl(resource.second)
-                                }
-
-                            } ?: run {
-                                currentFragment?.webView?.loadUrl(resource.second)
-                            }
-                        } else {
-                            resourcePager.currentItem = resource.first
-                        }
-                        break
-                    }
-                } else {
-                    resource as Triple<Int, String, String>
-                    if (resource.second.endsWith(href) || resource.third.endsWith(href)) {
-                        resourcePager.currentItem = resource.first
-                        break
-                    }
-                }
-            }
-        }
-
-        resourcePager.adapter = adapter
-
-        if (publication.metadata.presentation.layout == EpubLayout.REFLOWABLE) {
-            setCurrent(resourcesSingle)
-        } else {
-
-            when (preferences.getInt(COLUMN_COUNT_REF, 0)) {
-                1 -> {
-                    setCurrent(resourcesSingle)
-                }
-                2 -> {
-                    setCurrent(resourcesDouble)
-                }
-                else -> {
-                    // TODO based on device
-                    // TODO decide if 1 page or 2 page
-                    setCurrent(resourcesSingle)
-                }
-            }
-        }
-
-        if (supportActionBar!!.isShowing && allowToggleActionBar) {
-            resourcePager.systemUiVisibility = (View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    or View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                    or View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-                    or View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                    or View.SYSTEM_UI_FLAG_FULLSCREEN // hide status bar
-                    or View.SYSTEM_UI_FLAG_IMMERSIVE)
-        }
-
-        return true
-    }
-
-    override fun go(link: Link, animated: Boolean, completion: () -> Unit): Boolean {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun goForward(animated: Boolean, completion: () -> Unit): Boolean {
-        launch {
-            if (resourcePager.currentItem < resourcePager.adapter!!.count - 1) {
-
-                resourcePager.setCurrentItem(resourcePager.currentItem + 1, animated)
-
-                if (layoutDirectionIsRTL() || publication.contentLayout.readingProgression == ReadingProgression.RTL) {
-                    // The view has RTL layout
-                    currentFragment?.webView?.apply {
-                        progression = 1.0
-                        setCurrentItem(numPages - 1, false)
-                    }
-                } else {
-                    // The view has LTR layout
-                    currentFragment?.webView?.apply {
-                        progression = 0.0
-                        setCurrentItem(0, false)
-                    }
-                }
-            }
-        }
-        return true
-    }
-
-    override fun goBackward(animated: Boolean, completion: () -> Unit): Boolean {
-        launch {
-            if (resourcePager.currentItem > 0) {
-
-                resourcePager.setCurrentItem(resourcePager.currentItem - 1, animated)
-
-                if (layoutDirectionIsRTL() || publication.contentLayout.readingProgression == ReadingProgression.RTL) {
-                    // The view has RTL layout
-                    currentFragment?.webView?.apply {
-                        progression = 0.0
-                        setCurrentItem(0, false)
-                    }
-                } else {
-                    // The view has LTR layout
-                    currentFragment?.webView?.apply {
-                        progression = 1.0
-                        setCurrentItem(numPages - 1, false)
-                    }
-                }
-            }
-        }
-        return true
-    }
-
-    override val readingProgression: ReadingProgression
-        get() = TODO("not implemented") //To change initializer of created properties use File | Settings | File Templates.
-
-    override fun goLeft(animated: Boolean, completion: () -> Unit): Boolean {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
-    override fun goRight(animated: Boolean, completion: () -> Unit): Boolean {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-
     /**
      * Context of this scope.
      */
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Main
-
 
     override lateinit var preferences: SharedPreferences
     override lateinit var resourcePager: R2ViewPager
@@ -228,139 +61,27 @@ open class R2EpubActivity : AppCompatActivity(), IR2Activity, IR2Selectable, IR2
 
     override var allowToggleActionBar = true
 
-    private lateinit var resourcesSingle: ArrayList<Pair<Int, String>>
-    private lateinit var resourcesDouble: ArrayList<Triple<Int, String, String>>
 
-    var currentPagerPosition: Int = 0
     lateinit var adapter: R2PagerAdapter
 
     protected var navigatorDelegate: NavigatorDelegate? = null
 
+    @OptIn(FragmentNavigator::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_r2_viewpager)
 
         preferences = getSharedPreferences("org.readium.r2.settings", Context.MODE_PRIVATE)
-        resourcePager = findViewById(R.id.resourcePager)
-        resourcePager.type = Publication.TYPE.EPUB
-
-        resourcesSingle = ArrayList()
-        resourcesDouble = ArrayList()
 
         publication = intent.getPublication(this)
-        positions = runBlocking { publication.positions() }
         publicationPath = intent.getStringExtra("publicationPath") ?: throw Exception("publicationPath required")
         publicationFileName = intent.getStringExtra("publicationFileName") ?: throw Exception("publicationFileName required")
         publicationIdentifier = publication.metadata.identifier!!
 
-        title = null
+        supportFragmentManager.fragmentFactory = NavigatorFragmentFactory(publication, publicationPath, publicationFileName)
 
-        val port = preferences.getString("$publicationIdentifier-publicationPort", 0.toString())?.toInt()
+        setContentView(R.layout.activity_r2_epub)
 
-        // TODO needs work, currently showing two resources for fxl, needs to understand which two resources, left & right, or only right etc.
-        var doublePageIndex = 0
-        var doublePageLeft = ""
-        var doublePageRight = ""
-        var resourceIndexDouble = 0
-
-        for ((resourceIndexSingle, spineItem) in publication.readingOrder.withIndex()) {
-            val uri: String = if (publicationPath.toUri().isAbsolute) {
-                if (spineItem.href.toUri().isAbsolute) {
-                    spineItem.href
-                } else {
-                    getAbsolute(spineItem.href, publicationPath)
-                }
-            } else {
-                Publication.localUrlOf(filename = publicationFileName, port = port ?: 0, href = spineItem.href)
-            }
-            resourcesSingle.add(Pair(resourceIndexSingle, uri))
-
-            // add first page to the right,
-            if (resourceIndexDouble == 0) {
-                doublePageLeft = ""
-                doublePageRight = uri
-                resourcesDouble.add(Triple(resourceIndexDouble, doublePageLeft, doublePageRight))
-                resourceIndexDouble++
-            } else {
-                // add double pages, left & right
-                if (doublePageIndex == 0) {
-                    doublePageLeft = uri
-                    doublePageIndex = 1
-                } else {
-                    doublePageRight = uri
-                    doublePageIndex = 0
-                    resourcesDouble.add(Triple(resourceIndexDouble, doublePageLeft, doublePageRight))
-                    resourceIndexDouble++
-                }
-            }
-        }
-        // add last page if there is only a left page remaining
-        if (doublePageIndex == 1) {
-            doublePageIndex = 0
-            resourcesDouble.add(Triple(resourceIndexDouble, doublePageLeft, ""))
-        }
-
-
-        if (publication.metadata.presentation.layout == EpubLayout.REFLOWABLE) {
-            adapter = R2PagerAdapter(supportFragmentManager, resourcesSingle, publication.metadata.title, Publication.TYPE.EPUB, publicationPath)
-            resourcePager.type = Publication.TYPE.EPUB
-        } else {
-            resourcePager.type = Publication.TYPE.FXL
-            adapter = when (preferences.getInt(COLUMN_COUNT_REF, 0)) {
-                1 -> {
-                    R2PagerAdapter(supportFragmentManager, resourcesSingle, publication.metadata.title, Publication.TYPE.FXL, publicationPath)
-                }
-                2 -> {
-                    R2PagerAdapter(supportFragmentManager, resourcesDouble, publication.metadata.title, Publication.TYPE.FXL, publicationPath)
-                }
-                else -> {
-                    // TODO based on device
-                    // TODO decide if 1 page or 2 page
-                    R2PagerAdapter(supportFragmentManager, resourcesSingle, publication.metadata.title, Publication.TYPE.FXL, publicationPath)
-                }
-            }
-        }
-        resourcePager.adapter = adapter
-
-        resourcePager.direction = publication.contentLayout.readingProgression
-
-        if (publication.cssStyle == ReadingProgression.RTL.value) {
-            resourcePager.direction = ReadingProgression.RTL
-        }
-
-        resourcePager.addOnPageChangeListener(object : ViewPager.SimpleOnPageChangeListener() {
-
-            override fun onPageSelected(position: Int) {
-//                if (publication.metadata.presentation.layout == EpubLayout.REFLOWABLE) {
-//                    resourcePager.disableTouchEvents = true
-//                }
-                if (preferences.getBoolean(SCROLL_REF, false)) {
-                    if (currentPagerPosition < position) {
-                        // handle swipe LEFT
-                        currentFragment?.webView?.scrollToStart()
-                    } else if (currentPagerPosition > position) {
-                        // handle swipe RIGHT
-                        currentFragment?.webView?.scrollToEnd()
-                    }
-                } else {
-                    if (currentPagerPosition < position) {
-                        // handle swipe LEFT
-                        currentFragment?.webView?.setCurrentItem(0, false)
-                    } else if (currentPagerPosition > position) {
-                        // handle swipe RIGHT
-                        currentFragment?.webView?.apply {
-                            setCurrentItem(numPages - 1, false)
-                        }
-                    }
-                }
-                currentPagerPosition = position // Update current position
-
-                notifyCurrentLocation()
-            }
-
-        })
-
-
+        resourcePager = findViewById<View>(R.id.epub_navigator).rootView.findViewById(R.id.resourcePager)
     }
 
     override fun onDestroy() {
@@ -371,22 +92,11 @@ open class R2EpubActivity : AppCompatActivity(), IR2Activity, IR2Selectable, IR2
 
     override fun onResume() {
         super.onResume()
-        notifyCurrentLocation()
     }
 
     override fun onActionModeStarted(mode: ActionMode?) {
         mode?.menu?.clear()
         super.onActionModeStarted(mode)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == 2 && resultCode == Activity.RESULT_OK) {
-            val locator = data?.getParcelableExtra("locator") as? Locator
-            if (locator != null) {
-                go(locator)
-            }
-        }
-        super.onActivityResult(requestCode, resultCode, data)
     }
 
     override fun toggleActionBar() {
@@ -414,37 +124,35 @@ open class R2EpubActivity : AppCompatActivity(), IR2Activity, IR2Selectable, IR2
     private val currentFragment: R2EpubPageFragment? get() =
         r2PagerAdapter.mFragments.get(r2PagerAdapter.getItemId(resourcePager.currentItem)) as? R2EpubPageFragment
 
-    override val currentLocator: LiveData<Locator?> get() = _currentLocator
-    private val _currentLocator = MutableLiveData<Locator?>(null)
+    override val readingProgression: ReadingProgression
+        get() = TODO("Not yet implemented")
 
-    /**
-     * While scrolling we receive a lot of new current locations, so we use a coroutine job to
-     * debounce the notification.
-     */
-    private var debounceLocationNotificationJob: Job? = null
-
-    private fun notifyCurrentLocation() {
-        val navigator = this
-        debounceLocationNotificationJob?.cancel()
-        debounceLocationNotificationJob = launch {
-            delay(100L)
-
-            val resource = publication.readingOrder[resourcePager.currentItem]
-            val progression = currentFragment?.webView?.progression ?: 0.0
-            val positions = publication.positionsByResource[resource.href]
-                ?: return@launch
-            val positionIndex = ceil(progression * (positions.size - 1)).toInt()
-            val locator = positions[positionIndex]
-                .copyWithLocations(progression = progression)
-
-            if (locator == currentLocator.value) {
-                return@launch
-            }
-
-            _currentLocator.postValue(locator)
-            navigatorDelegate?.locationDidChange(navigator = navigator, locator = locator)
-        }
+    override fun goLeft(animated: Boolean, completion: () -> Unit): Boolean {
+        TODO("Not yet implemented")
     }
+
+    override fun goRight(animated: Boolean, completion: () -> Unit): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    override val currentLocator: LiveData<Locator?> get() = _currentLocator
+    override fun go(locator: Locator, animated: Boolean, completion: () -> Unit): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    override fun go(link: Link, animated: Boolean, completion: () -> Unit): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    override fun goForward(animated: Boolean, completion: () -> Unit): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    override fun goBackward(animated: Boolean, completion: () -> Unit): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    private val _currentLocator = MutableLiveData<Locator?>(null)
 
     override fun currentSelection(callback: (Locator?) -> Unit) {
         currentFragment?.webView?.getCurrentSelectionInfo {
@@ -561,12 +269,6 @@ open class R2EpubActivity : AppCompatActivity(), IR2Activity, IR2Selectable, IR2
                 }
             }
         }
-
     }
-
-    override fun onPageLoaded() {
-        super.onPageLoaded()
-    }
-
 }
